Rename D2Client function folder
