+++ SGD2FreeResolution/src/sgd2mapi_extension/game_function/d2client/d2client_draw_screens.cpp
