+++ SGD2FreeResolution/src/sgd2mapi_extension/game_function/d2client/d2client_draw_screens.hpp
@@ -43,8 +43,8 @@
  *  work.
  */
 
-#ifndef SGD2FR_SGD2MAPI_EXTENSION_GAME_FUNCTION_D2CLIENT_FUNCTION_D2CLIENT_DRAW_SCREENS_HPP_
-#define SGD2FR_SGD2MAPI_EXTENSION_GAME_FUNCTION_D2CLIENT_FUNCTION_D2CLIENT_DRAW_SCREENS_HPP_
+#ifndef SGD2FR_SGD2MAPI_EXTENSION_GAME_FUNCTION_D2CLIENT_D2CLIENT_DRAW_SCREENS_HPP_
+#define SGD2FR_SGD2MAPI_EXTENSION_GAME_FUNCTION_D2CLIENT_D2CLIENT_DRAW_SCREENS_HPP_
 
 namespace d2 {
 namespace d2client {
@@ -54,4 +54,4 @@ void DrawScreens();
 } // namespace d2client
 } // namespace d2
 
-#endif /* SGD2FR_SGD2MAPI_EXTENSION_GAME_FUNCTION_D2CLIENT_FUNCTION_D2CLIENT_DRAW_SCREENS_HPP_ */
+#endif /* SGD2FR_SGD2MAPI_EXTENSION_GAME_FUNCTION_D2CLIENT_D2CLIENT_DRAW_SCREENS_HPP_ */
